+++ wright/src/grammar/ast/literal.rs
@@ -1,0 +1,55 @@
+use crate::grammar::model::Fragment;
+
+/// Numerical literal in wright source code.
+/// i.e. `10`, `0xCa1a0`, `0b0101_0101`, `100_000`
+#[derive(Copy, Clone, Debug)]
+pub struct NumLit<'s> {
+    /// Associated fragment of source code.
+    pub frag: Fragment<'s>,
+    /// Represented value.
+    pub inner: u128,
+}
+
+/// Character literal in wright source code.
+/// i.e `'a', '\n', '\u{01f441}', '\x00', '♦'`
+/// see [this page](https://doc.rust-lang.org/reference/tokens.html#ascii-escapes) for escape
+/// information.
+#[derive(Copy, Clone, Debug)]
+pub struct CharLit<'s> {
+    /// Associated fragment of source code.
+    pub frag: Fragment<'s>,
+    /// Represented Value.
+    pub inner: char,
+}
+
+/// String literal in wright source code.
+/// i.e. `"hello world", "with \n newline \n escapes"`
+/// The parser for string literals also includes all the escape characters
+/// [here](https://doc.rust-lang.org/reference/tokens.html#ascii-escapes).
+/// Raw-strings and Byte-strings (like those in rust) are not currently
+/// supported but may be added in the future.
+#[derive(Clone, Debug)]
+pub struct StringLit<'s> {
+    /// Associated fragment of source code.
+    pub frag: Fragment<'s>,
+    /// Represented string value. (not a reference into source code because
+    /// source code may contain escaped characters.)
+    pub inner: String,
+}
+
+/// Boolean literal in wright source code.
+/// i.e. `true`, `false`.
+#[derive(Copy, Clone, Debug)]
+pub struct BooleanLit<'s> {
+    /// Associated fragment in source code.
+    pub frag: Fragment<'s>,
+    /// Represented value.
+    pub inner: bool,
+}
+
+/// `self` literal in wright source code.
+#[derive(Copy, Clone, Debug)]
+pub struct SelfLit<'s> {
+    /// Associated fragment in source code.
+    pub frag: Fragment<'s>,
+}
