+++ wright/src/grammar/ast/statement.rs
@@ -1,0 +1,18 @@
+use crate::grammar::model::Fragment;
+use crate::grammar::ast::Expression;
+
+/// An expression whose results aren't used.
+#[derive(Clone, Debug)]
+pub struct ExpressionStatement<'s> {
+    /// Associated Fragment in source code.
+    pub frag: Fragment<'s>,
+    /// The expression.
+    pub inner: Box<Expressionn<'s>>,
+}
+
+/// A statement in wright source code.
+#[allow(missing_docs)]
+#[derive(Clone, Debug)]
+pub enum Statement<'s> {
+    ExpressionStatement(ExpressionStatement<'s>),
+}
