+++ wright/src/grammar/ast/types.rs
@@ -1,0 +1,12 @@
+
+/// A wright primitive type.
+pub enum PrimitiveTypeVariant {
+
+}
+
+/// A type in source code.
+#[allow(missing_docs)]
+#[derive(Clone, Debug)]
+pub enum Type<'s> {
+
+}
